Renamed excel(x) to xls(x) parsers to follow general pattern (#61)
